Diagnostics: SCSI read/tiling probes + WREG cycle gate; JRN-SCSI-027
Instruments (not product behaviour) that together exonerated the whole
I/O stack by evidence:

- Ncr53C810 N810-CMD trace now carries the CDB-decoded LBA/count, an
  FNV-1a/64 checksum of the returned payload, its head/tail bytes, and
  the full CDB.  Data-OUT commands take their payload from the
  delivered buffer, since handleCommand() zeroes dataTransferred on
  entry -- that fix is what let MODE SELECT's block descriptor testify
  its own block length (0x000200).
- NEW N810-MOVE tiling probe on the HBA->guest-RAM hop, two-tier per
  house rule: EMULATR_BRINGUP_PROBES compile guard outside,
  EMULATR_SCSI_MOVE_PROBE env key inside (the ITBPROBE shape).  It
  logs the guest PHYSICAL address per MOVE, which also evidences that
  delivery is a direct PA poke with no bus-master window translation.
- NEW tools/scsi_read_diff.py: re-reads each traced LBA from the
  backing image, recomputes the same checksum, and names the
  divergence shape (truncation vs pad vs wrong offset).

- PipelineDriver: EMULATR_DIAG_WREG now honours EMULATR_DIAG_CYCLO/
  CYCHI.  The cycle window gated only the DIAG-PC path, so every
  register probe was implicitly console-era-only -- its cap filling a
  billion cycles before any OS-era target.  Same gate, same defaults;
  existing recipes unaffected.

Results (JRN-SCSI-027): 68/68 READ payloads byte-match dka0.vdisk (an
image Charon boots into OpenVMS), 47/47 commands tile guest RAM
exactly once with contiguous PAs, zero padding in the SYSBOOT window,
and geometry closed by the driver's own descriptor.  So
%SYSBOOT-F-LDFAIL = %LOADER-E-BADIMGOFF means correct bytes, correctly
placed, wrong offset computed from them.  Loader anchor captured by
value: pc 0x42790 LDQ r0,-0x10(r27) loads the status from a linkage
cell.  Next session: retire window upstream of it, diff the ISD field
loads against the host-side image parse.

Gate: full suite 502/505 (the 3 pre-existing drift failures only);
DS20 dka0 boot verified to the current frontier with all of this in.

+++ tests/palBoxLib/test_palentries.cpp
@@ -974,8 +974,8 @@ TEST_CASE("palBox::execMtprVptb_vms -- updates cpu.vptb, both IPRs, AND PT__VPTB
     // PHYSICAL, like the .mar's hw_stq/p.  Without S_PhysAddr the Mbox walks
     // p_temp as a VA, the store double-misses, and the machine halts 0x0A at
     // the MTPR site -- the failure mode seen on this fix's first cut.
-    CHECK((r.semFlags & GrainSem::S_PhysAddr) != 0);
-    CHECK((r.semFlags & GrainSem::S_Store) != 0);
+    CHECK(static_cast<uint64_t>(r.semFlags & GrainSem::S_PhysAddr) != 0u);
+    CHECK(static_cast<uint64_t>(r.semFlags & GrainSem::S_Store) != 0u);
 }
 
 TEST_CASE("palBox::execMtprVptb_vms -- an unusable p_temp FAULTS, never skips")
